Set AccessibilityTitle option properly
Updated to better handle a11y, as follows:
- For option controls that have a Title property (e.g. NSButton normally has a
Title, which is the button label), don't set anything for a11y, so that the title
is read by voice over.
- For option controls that don't have a Title property (e.g. NSTextField), then
set the AccessibilityTitle property to be the label for the control (if there is one).

This handles most of what we want for a11y. There are still a few things missing here,
but this is the bulk of the needed support.